requirements.txt
beautifulsoup4==4.13.4
mariadb==1.1.12
python-dotenv==1.1.0
Requests==2.32.3
